+++ benchmark/pack_matmul/pack_matmul_registry.cpp
@@ -28,6 +28,7 @@
 #endif  // __GNUC__
 
 // Matrix multiplication micro-kernels with associated LHS packing micro-kernels.
+#include "kai/ukernels/matmul/kai_matmul.h"
 #include "kai/ukernels/matmul/matmul_clamp_bf16_qai8dxp_qsi4c32p/kai_matmul_clamp_bf16_qai8dxp1x8_qsi4c32p4x8_1x4_neon_dotprod.h"
 #include "kai/ukernels/matmul/matmul_clamp_bf16_qai8dxp_qsi4c32p/kai_matmul_clamp_bf16_qai8dxp4x8_qsi4c32p4x8_16x4_neon_i8mm.h"
 #include "kai/ukernels/matmul/matmul_clamp_bf16_qai8dxp_qsi4cxp/kai_matmul_clamp_bf16_qai8dxp1x8_qsi4cxp8x8_1x8_neon_dotprod.h"
@@ -673,6 +674,94 @@ inline const PackMatMulEntry kai_matmul_clamp_f32_qai8dxp1x4_qsi4c32p4vlx4_1x4vl
         kai_run_lhs_quant_pack_qai8dxp_f32_as_void, kai_run_matmul_clamp_f32_qai8dxp1x4_qsi4c32p4vlx4_1x4vl_sme2_dot>,
 };
 
+size_t kai_get_mr_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa(void) {
+    const kai_matmul_uker_config config{};
+    return kai_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa().get_step(&config).m;
+}
+
+size_t kai_get_kr_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa(void) {
+    return 8;
+}
+
+size_t kai_get_sr_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa(void) {
+    return 2;
+}
+
+size_t kai_get_lhs_packed_offset_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa(
+    size_t m_idx, size_t k) {
+    const auto api = kai_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa();
+    const kai_matmul_uker_config config{};
+    const kai_matmul_uker_lhs_dim_args shape{0, k};
+    const auto stride = api.get_lhs_stride(&config, &shape);
+    const kai_matmul_uker_lhs_dim_args index{m_idx, 0};
+    return api.get_lhs_offset(&config, &index, &stride);
+}
+
+inline const PackMatMulEntry kai_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa_lhs_pack_entry{
+    .benchmark_name =
+        "kai_pack_matmul/kai_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa/"
+        "kai_lhs_quant_pack_qai8dxp_f32",
+    .matmul_name = "kai_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa",
+    .lhs_type = DataType::FP32,
+    .dst_type = DataType::FP32,
+    .matmul_op = PackMatMulOp::GEMM,
+    .needs_block_size = true,
+    .is_cpu_supported = test::cpu_has_sme,
+    .get_mr = kai_get_mr_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa,
+    .get_kr = kai_get_kr_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa,
+    .get_sr = kai_get_sr_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa,
+    .get_lhs_offset = kai_get_lhs_offset_lhs_quant_pack_qai8dxp_f32,
+    .get_lhs_packed_offset = kai_get_lhs_packed_offset_lhs_quant_pack_qai8dxp_f32,
+    .get_lhs_packed_size = kai_get_lhs_packed_size_lhs_quant_pack_qai8dxp_f32,
+    .get_matmul_lhs_packed_offset =
+        kai_get_lhs_packed_offset_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa,
+    .run_pack_matmul = run_pack_matmul_uker_api<
+        kai_run_lhs_quant_pack_qai8dxp_f32_as_void, kai_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa>,
+};
+
+size_t kai_get_mr_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot(void) {
+    const kai_matmul_uker_config config{};
+    return kai_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot().get_step(&config).m;
+}
+
+size_t kai_get_kr_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot(void) {
+    return 8;
+}
+
+size_t kai_get_sr_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot(void) {
+    return 2;
+}
+
+size_t kai_get_lhs_packed_offset_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot(size_t m_idx, size_t k) {
+    const auto api = kai_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot();
+    const kai_matmul_uker_config config{};
+    const kai_matmul_uker_lhs_dim_args shape{0, k};
+    const auto stride = api.get_lhs_stride(&config, &shape);
+    const kai_matmul_uker_lhs_dim_args index{m_idx, 0};
+    return api.get_lhs_offset(&config, &index, &stride);
+}
+
+inline const PackMatMulEntry kai_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot_lhs_pack_entry{
+    .benchmark_name =
+        "kai_pack_matmul/kai_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot/"
+        "kai_lhs_quant_pack_qai8dxp_f32",
+    .matmul_name = "kai_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot",
+    .lhs_type = DataType::FP32,
+    .dst_type = DataType::FP32,
+    .matmul_op = PackMatMulOp::GEMV,
+    .needs_block_size = true,
+    .is_cpu_supported = test::cpu_has_sme,
+    .get_mr = kai_get_mr_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot,
+    .get_kr = kai_get_kr_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot,
+    .get_sr = kai_get_sr_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot,
+    .get_lhs_offset = kai_get_lhs_offset_lhs_quant_pack_qai8dxp_f32,
+    .get_lhs_packed_offset = kai_get_lhs_packed_offset_lhs_quant_pack_qai8dxp_f32,
+    .get_lhs_packed_size = kai_get_lhs_packed_size_lhs_quant_pack_qai8dxp_f32,
+    .get_matmul_lhs_packed_offset = kai_get_lhs_packed_offset_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot,
+    .run_pack_matmul = run_pack_matmul_uker_api<
+        kai_run_lhs_quant_pack_qai8dxp_f32_as_void, kai_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot>,
+};
+
 inline const PackMatMulEntry kai_matmul_clamp_bf16_qai8dxp1x8_qsi4c32p4x8_1x4_neon_dotprod_lhs_pack_entry{
     .benchmark_name =
         "kai_pack_matmul/kai_matmul_clamp_bf16_qai8dxp1x8_qsi4c32p4x8_1x4_neon_dotprod/"
@@ -726,7 +815,7 @@ PackMatMulRegistryEntry RegisterPackMatMulBenchmarkEntry(const PackMatMulEntry& 
 ///
 /// @return Registered pack matmul benchmarks.
 const auto& get_pack_matmul_benchmarks() {
-    static const std::array<PackMatMulRegistryEntry, 30> pack_matmul_entries{
+    static const std::array<PackMatMulRegistryEntry, 32> pack_matmul_entries{
         {
             RegisterPackMatMulBenchmarkEntry(kai_matmul_clamp_f16_bf16p8x4_bf16p12x4b_8x12_neon_mmla_lhs_pack_entry),
             RegisterPackMatMulBenchmarkEntry(kai_matmul_clamp_f16_f16p2vlx2_f16p2vlx2_2vlx2vl_sme2_mopa_lhs_pack_entry),
@@ -739,6 +828,10 @@ const auto& get_pack_matmul_benchmarks() {
                 kai_matmul_clamp_f32_qai8dxp1x4_qsi4c32p4vlx4_1x4vl_sme2_dot_lhs_pack_entry),
             RegisterPackMatMulBenchmarkEntry(
                 kai_matmul_clamp_f32_qai8dxp1vlx4_qsi4c32p4vlx4_1vlx4vl_sme2_mopa_lhs_pack_entry),
+            RegisterPackMatMulBenchmarkEntry(
+                kai_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot_lhs_pack_entry),
+            RegisterPackMatMulBenchmarkEntry(
+                kai_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa_lhs_pack_entry),
             RegisterPackMatMulBenchmarkEntry(
                 kai_matmul_clamp_f32_qai8dxp1vlx8_qsi4cxp4vlx8_1vlx4vl_sme2_mopa_lhs_pack_entry),
             RegisterPackMatMulBenchmarkEntry(