+++ kai/ukernels/matmul/kai_matmul.h
@@ -406,6 +406,58 @@ struct kai_matmul_uker_api kai_matmul_clamp_f32_f16p4vsx2_qai4c32p16vsx4s1s0sf16
 /// @return The micro-kernel API.
 struct kai_matmul_uker_api kai_matmul_clamp_f32_qsi8d32p1x4_qai4c32p16vsx4s1s0sf16_1x16vs_sme2_dot(void);
 
+/// Single-precision floating-point matrix multiplication with statically-blockwise quantized unsigned INT4 RHS,
+/// using SME MOPA instruction.
+///
+/// Required CPU features:
+///   * FEAT_SME
+///
+/// Configuration parameters:
+///   * bl - The block length, in number of K values per block. Must be a multiple of 32.
+///
+/// Required operands:
+///   * dst
+///   * lhs - Dynamically quantized asymmetric per-row (qai8dxp) LHS, packed with @ref
+///     kai_lhs_quant_pack_qai8dxp_f32.
+///   * rhs - Packed RHS matrix, obtained with @ref kai_matmul_pack_rhs_kxn_qsi4c32p16vsx4s4s0_qsu4c32_f32_bf16_sme or
+///   @ref
+///     kai_matmul_pack_rhs_nxk_qsi4c32p16vsx4s4s0_qsu4c32_f32_bf16_sme.
+///
+/// Optional arguments:
+///   * clamp - F32 output clamp values if KAI_MATMUL_UKER_FLAGS_ARGS_CLAMP flag is set.
+///
+/// Supported flags:
+///   * KAI_MATMUL_UKER_FLAGS_ARGS_CLAMP - Clamp output data.
+///
+/// @return The micro-kernel API.
+struct kai_matmul_uker_api kai_matmul_clamp_f32_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa(void);
+
+/// Single-precision floating-point vector-matrix multiplication with statically-blockwise quantized unsigned INT4
+/// RHS, using SME DOT instruction.
+///
+/// Required CPU features:
+///   * FEAT_SME
+///
+/// Configuration parameters:
+///   * bl - The block length, in number of K values per block. Must be a multiple of 32.
+///
+/// Required operands:
+///   * dst
+///   * lhs - Dynamically quantized asymmetric per-row (qai8dxp) LHS, packed with @ref
+///     kai_lhs_quant_pack_qai8dxp_f32.
+///   * rhs - Packed RHS matrix, obtained with @ref kai_matmul_pack_rhs_kxn_qsi4c32p16vsx4s4s0_qsu4c32_f32_bf16_sme or
+///   @ref
+///     kai_matmul_pack_rhs_nxk_qsi4c32p16vsx4s4s0_qsu4c32_f32_bf16_sme.
+///
+/// Optional arguments:
+///   * clamp - F32 output clamp values if KAI_MATMUL_UKER_FLAGS_ARGS_CLAMP flag is set.
+///
+/// Supported flags:
+///   * KAI_MATMUL_UKER_FLAGS_ARGS_CLAMP - Clamp output data.
+///
+/// @return The micro-kernel API.
+struct kai_matmul_uker_api kai_matmul_clamp_f32_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot(void);
+
 #ifdef __cplusplus
 }  // extern "C"
 #endif