+++ test/nextgen/operators/matmul/matmul_slots.hpp
@@ -48,13 +48,15 @@ enum class MatMulSlot : size_t {
     RHS_PACKED,      ///< Packed RHS.
     RHS_PACKED_IMP,  ///< Packed RHS from micro-kernel.
 
-    RHS_T_DATA,            ///< Transposed RHS data.
-    RHS_T_CVT_DATA,        ///< Transposed RHS data after conversion.
-    RHS_T_QDATA,           ///< Transposed RHS data after quantization.
-    RHS_T_QDATA_SIGN,      ///< Transposed RHS data after quantization with opposite signedness.
-    RHS_T_QDATA_SIGN_T,    ///< Transposed RHS_T_QDATA_SIGN.
-    RHS_T_QDATA_SIGN_SUM,  ///< Row sum of transposed RHS after quantization with opposite signedness.
-    RHS_T_QSCALE,          ///< Transposed RHS quantization scale.
+    RHS_T_DATA,                   ///< Transposed RHS data.
+    RHS_T_CVT_DATA,               ///< Transposed RHS data after conversion.
+    RHS_T_QDATA,                  ///< Transposed RHS data after quantization.
+    RHS_T_QDATA_SIGN,             ///< Transposed RHS data after quantization with opposite signedness.
+    RHS_T_QDATA_SIGN_T,           ///< Transposed RHS_T_QDATA_SIGN.
+    RHS_T_QDATA_SIGN_SUM,         ///< Row sum of transposed RHS after quantization with opposite signedness.
+    RHS_T_QSCALE,                 ///< Transposed RHS quantization scale.
+    RHS_T_QSCALE_RESCALED,        ///< Transposed RHS quantization scale, rescaled by `1/16`.
+    RHS_T_QDATA_SIGN_SUM_SCALED,  ///< Row sum of transposed RHS after quantization, weighted by the block scale.
     RHS_T_QSCALE_MUL_LHS_QSCALE_DIV_DST_QSCALE,  ///< Static Int8 RHS packed scale component.
     RHS_T_QZP,                                   ///< Transposed RHS quantization zero-point.