+++ test/nextgen/operators/matmul/pack_lhs/matmul_pack_lhs_wrapper_registry.cpp
@@ -12,6 +12,7 @@
 #include <string>
 #include <string_view>
 
+#include "kai/ukernels/matmul/kai_matmul.h"
 #include "kai/ukernels/matmul/kai_matmul_pack_lhs.h"
 #include "kai/ukernels/matmul/matmul_clamp_f32_f32p_f32p/kai_matmul_clamp_f32_f32p2vlx1_f32p2vlx1biasf32_sme2_mopa.h"
 #include "kai/ukernels/matmul/matmul_clamp_f32_qai8dxp_qsi4cxp/kai_matmul_clamp_f32_qai8dxp1vlx4_qsi4cxp4vlx4_1vlx4vl_sme_mopa.h"
@@ -190,6 +191,10 @@ std::unique_ptr<KernelWrapper<MatShape>> create_matmul_matmul_pack_lhs_mxk_qsi8d
     );
 }
 
+std::unique_ptr<KernelWrapper<MatShape>> create_matmul_lhs_quant_pack_qai8dxp4vsx4_f32() {
+    return create_matmul_lhs_quant_pack_qai8dxp_f32("4vsx4", 4 * get_sme_vector_scale(), 4);
+}
+
 std::unique_ptr<KernelWrapper<MatShape>> create_matmul_lhs_pack_f32p2vlx1_f32_sme() {
     return std::make_unique<MatMulPackLhsFpWrapper>(
         "create_matmul_lhs_pack_f32p2vlx1_f32_sme",
@@ -287,6 +292,29 @@ bool is_shape_suitable_lhs_f32p2vlx1_f32p2vlx1biasf32_sme2_mopa(
 
     const size_t mr = kai_get_mr_matmul_clamp_f32_f32p2vlx1_f32p2vlx1biasf32_sme2_mopa();
     const size_t lhs_m_step = kai_get_m_step_lhs_pack_f32p2vlx1_f32_sme(mr);
+
+    return portion_non_empty(shape_m, shape_k, lhs_m_step, shape_k, portion);
+}
+
+bool is_shape_suitable_lhs_qai8dxp1x4_qsi4c32p16vsx4_1x16vs_sme_dot(
+    size_t shape_m, [[maybe_unused]] size_t shape_n, size_t shape_k, const MatrixPortion& portion) {
+    if (shape_m == 0 || shape_k == 0 || (shape_k % 32) != 0) {
+        return false;
+    }
+
+    const size_t lhs_m_step = kai_get_m_step_lhs_quant_pack_qai8dxp_f32(1);
+
+    return portion_non_empty(shape_m, shape_k, lhs_m_step, shape_k, portion);
+}
+
+bool is_shape_suitable_lhs_qai8dxp4vsx4_qsi4c32p16vsx4_4vsx16vs_sme_mopa(
+    size_t shape_m, [[maybe_unused]] size_t shape_n, size_t shape_k, const MatrixPortion& portion) {
+    if (shape_m == 0 || shape_k == 0 || (shape_k % 32) != 0) {
+        return false;
+    }
+
+    const size_t mr = 4 * get_sme_vector_scale();
+    const size_t lhs_m_step = kai_get_m_step_lhs_quant_pack_qai8dxp_f32(mr);
 
     return portion_non_empty(shape_m, shape_k, lhs_m_step, shape_k, portion);
 }