+++ test/nextgen/operators/matmul/pack_rhs/matmul_pack_rhs_ukerapi_t_wrapper.hpp
@@ -24,10 +24,10 @@ public:
         const Poly<Format>& src_bias_format, const Poly<Format>& dst_format,
         MatMulUkerApiBiasDeliveryStage bias_delivery_stage, MatMulPackRhsOperandSlots operand_slots = {},
         std::vector<MatMulSlot> reference_component_slots = {MatMulSlot::RHS_T_DATA},
-        MatMulSlot run_rhs_slot = MatMulSlot::RHS_T_DATA) :
+        MatMulSlot run_rhs_slot = MatMulSlot::RHS_T_DATA, kai_matmul_pack_rhs_uker_format_config format = {}) :
         MatMulPackRhsUkerApiCommon(
             name, run_rhs_slot, RhsLayout::NxK, api, src_data_format, src_bias_format, dst_format, bias_delivery_stage,
-            operand_slots, std::move(reference_component_slots)) {
+            operand_slots, std::move(reference_component_slots), format) {
     }
 };